src_requirements.txt
streamlit==1.45.1
psycopg[binary]==3.2.9
databricks-sdk>=0.55.0
pandas>=2.0.0
openai>=1.0.0
mlflow[databricks]>=3.1.0
langgraph>=0.2.0
langchain-core>=0.3.0
